MK-07: Refactor of VideoPlayer
