+++ package.json
@@ -38,7 +38,7 @@
     "react-native-svg": "^15.10.1",
     "react-native-vector-icons": "^10.2.0",
     "react-native-video": "^6.8.2",
-    "react-native-webview": "^13.12.5" 
+    "react-native-webview": "^13.12.5"
   },
   "devDependencies": {
     "@babel/core": "^7.25.2",
